+++ docs/HISTORY.md
@@ -11,6 +11,7 @@ None
 - Fixed a cosmetic error on admin stream disconnect
 - Fixed spurious entry in service dictionary causing dashboard `__name`` errors
 - Added setter for setter for `pin_apps`
+- Passing through keyword arguments to `render_template`
 - Fixed a bug in the error text for apps in a custom app_dir
 
 **Breaking Changes**